# Faster merge functions