Some optimisations on the `merge_{hi,lo}` functions