@Kerl13 2021-02-16
## Before

```
================================== [ Unif ] ==================================
length     stdlib    timsort   speedup    lg(n!)    stdlib   timsort   speedup
------------------------------------------------------------------------------
 30000   0.007039   0.010258    -45.7%    402909    409951    447116     -9.1%
 60000   0.014751   0.021221    -43.9%    865809    879910    955407     -8.6%
 90000   0.023624   0.031385    -32.9%   1351355   1374063   1470859     -7.0%
120000   0.031248   0.045396    -45.3%   1851608   1879818   2056774     -9.4%
150000   0.040328   0.056562    -40.3%   2362797   2406163   2595284     -7.9%
------------------------------------------------------------------------------

================================= [ 5-Runs ] =================================
length     stdlib    timsort   speedup    lg(n!)    stdlib   timsort   speedup
------------------------------------------------------------------------------
 30000   0.004255   0.002176     48.9%    402909    260971     97146     62.8%
 60000   0.008782   0.003431     60.9%    865809    552184    191963     65.2%
 90000   0.013131   0.004885     62.8%   1351355    864474    294032     66.0%
120000   0.018827   0.007167     61.9%   1851608   1179318    387267     67.2%
150000   0.023528   0.008382     64.4%   2362797   1495840    488300     67.4%
------------------------------------------------------------------------------
```

## After 

```
================================== [ Unif ] ==================================
length     stdlib    timsort   speedup    lg(n!)    stdlib   timsort   speedup
------------------------------------------------------------------------------
 30000   0.007027   0.009505    -35.3%    402909    409951    447116     -9.1%
 60000   0.015158   0.019277    -27.2%    865809    879910    955407     -8.6%
 90000   0.023195   0.029690    -28.0%   1351355   1374063   1470859     -7.0%
120000   0.030921   0.042278    -36.7%   1851608   1879818   2056774     -9.4%
150000   0.040280   0.053288    -32.3%   2362797   2406163   2595284     -7.9%
------------------------------------------------------------------------------

================================= [ 5-Runs ] =================================
length     stdlib    timsort   speedup    lg(n!)    stdlib   timsort   speedup
------------------------------------------------------------------------------
 30000   0.004224   0.001736     58.9%    402909    260971     97146     62.8%
 60000   0.008755   0.002680     69.4%    865809    552184    191963     65.2%
 90000   0.013104   0.004630     64.7%   1351355    864474    294032     66.0%
120000   0.018372   0.006515     64.5%   1851608   1179318    387267     67.2%
150000   0.023722   0.007637     67.8%   2362797   1495840    488300     67.4%
------------------------------------------------------------------------------
```