@Kerl13 2021-02-17
I implemented one optimisation here, which reduces the number of memory accesses.

Before the patch, the `merge_{hi,lo}` functions would perform 4 `Array.get` per recursive call.
- 2 of them could easily be avoided (we were really accessing every element twice in a row)
- of the remaining 2 memory access, one could be reused at the next recursive call. So I changed the logic of the functions so that the **caller** makes the memory accesses rather than the callee, this allows to "pass along" to the next recursive call the value that will be reused

→ we now perform one memory access per recursive call
→ downside: it clutters the code a bit…