@Kerl13 2021-02-17
And with the last commit:

```
================================== [ Unif ] ==================================
length     stdlib    timsort   speedup    lg(n!)    stdlib   timsort   speedup
------------------------------------------------------------------------------
 30000   0.006782   0.008458    -24.7%    402909    409955    446234     -8.8%
 60000   0.014632   0.017702    -21.0%    865809    879918    956292     -8.7%
 90000   0.022037   0.027066    -22.8%   1351355   1374016   1472769     -7.2%
120000   0.030500   0.037782    -23.9%   1851608   1879819   2042806     -8.7%
150000   0.038565   0.048995    -27.0%   2362797   2406175   2590843     -7.7%
------------------------------------------------------------------------------

================================= [ 5-Runs ] =================================
length     stdlib    timsort   speedup    lg(n!)    stdlib   timsort   speedup
------------------------------------------------------------------------------
 30000   0.003933   0.001502     61.8%    402909    261693     97596     62.7%
 60000   0.008345   0.003069     63.2%    865809    553642    194825     64.8%
 90000   0.012528   0.004685     62.6%   1351355    856814    292144     65.9%
120000   0.017935   0.005648     68.5%   1851608   1169300    391936     66.5%
150000   0.022365   0.006656     70.2%   2362797   1499453    487716     67.5%
------------------------------------------------------------------------------
```